Add files via upload
Scratch file:  1) A banana is set to move with the mouse pointer, it changes color and stamps the canvas with that color repeatedly, leaving a trail of stamps as the mouse moves.  2) An apple changes size by 10 when the up arrow key is press and conversely changes size by -10 when the down arrow key is pressed.  Each time the apple is click on by the mouse clicker, a counter changes by an increase of one.  3) A birthday cake moves 10 to the right when it is clicked on by the mouse pointer.